requirements.txt
webdriver-manager
pytest
selenium
allure-pytest